feat: enable consul dns from nomad client nodes
getting consul dns and systemd-resolved to work nicely in docker is
a bit of a hack job. In docker /etc/resolv.conf can't be modified so we
clobber it with the /run/systemd/resolve/stub-resolv.conf config.

to do the clobbering, we are using a oneshot systemd service to do it on
boot.

I couldn't find an easy way to retain the docker dns setting for
127.0.0.11 so for now I have added it as a recursor to the consul dns
config. perhaps this could be fixed with some sort of network config,
but this seems to do the job for now.